+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## v0.4.1 (2026-02-25)
+
+### Added
+
+- Support for passing a custom alphabet (`.chars()`) to generate the captcha text.
+- Deferred text generation to `.build()` to allow flexible builder ordering.
+
 ## v0.4.0 (2026-02-25)
 
 ### Added